Move actual ImageMagick calling into separate file
This gets rid of the weird, brain-melting ouroboros of code that
recurses across threads and processes.
